Refactor booking and promotions components to improve type safety and remove unnecessary state management
